\name{CSurf-package}
\alias{CSurf-package}
\alias{CSurf}
\docType{package}
\title{
\packageTitle{CSurf}
}
\description{
\packageDescription{CSurf}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{CSurf}
\packageIndices{CSurf}
}
\author{
\packageAuthor{CSurf}

Maintainer: \packageMaintainer{CSurf}
}
\references{
~~ Literature or other references for background information ~~
}