% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I-SLAMM.R
\name{CSurf}
\alias{CSurf}
\title{CSurf}
\usage{
CSurf(
  y,
  X,
  f_hat0,
  h = 0.5,
  Tighten = FALSE,
  tol_out = 1e-04,
  maxiter_out = 10,
  m = 3,
  lam2rate = 1
)
}
\arguments{
\item{y}{A vector (length n) of the response for subgroup identification.}

\item{X}{A matrix (n x p) of the covariates for subgroup identification. The covariates are also used as potential thresholding variables for subgroup identification.}

\item{f_hat0}{A vector (length n) of the initialization for the threshold function (change surface).}

\item{Tighten}{A boolean indicating whether to include the tightening stage in optimization.}

\item{tol_out}{The tolerance value for optimization (stopping criterion).}

\item{maxiter_out}{The maximal number of iterations allowed for optimization.}

\item{m}{An integer indicating the power of the threshold function.}

\item{lam2rate}{The ratio between the two tunning parameters, lambda1 and lambda2.}
}
\value{
A list containing the following fields: 
\item{f_all.hat}{The estimated threshold functions. Each column represent one component function.}
\item{tau.hat}{The estimated change points over the threshold functions.}
\item{alp.hat}{The estimated regression coefficients.}
\item{coverged_out}{An boolean indicating whether the algorithm successfully converged.}
\item{iter_out}{The number of iterations it took to converge.}
\item{...}{Other estimation results.}
}
\description{
Change Surface (CSurf) regression method (additive model based) for nonlinear subgroup identification
}
\examples{
data(syn_data)
results <- CSurf(y = y, X = X, f_hat0 = f_hat0, h = .5, maxiter_out = 10, tol_out = 0.0001)
}
