% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/I-SLAMM.R
\name{pred_Fun}
\alias{pred_Fun}
\title{pred_Fun}
\usage{
pred_Fun(X, y, f, tau, alp)
}
\arguments{
\item{X}{A matrix (n x p) of the covariates. The covariates are also used as potential thresholding variables for subgroup identification.}

\item{y}{A vector (length n) of the true response.}

\item{f}{A vector (length n) the threshold function (change surface) obtained by CSurf.}

\item{tau}{The change points over the threshold functions obtained by CSurf.}

\item{alp}{The estimated regression coefficients obtained by CSurf.}
}
\value{
A list containing the following fields: 
\item{Yhat}{A vector of the predicted response.}
\item{mse}{The mean squared error (MSE) of response prediction.}
}
\description{
The function for prediction based on the Change Surface (CSurf) regression results
}
